requirements.txt
# Multi-Agent MCP System Requirements

# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
Alembic==1.12.1
redis==5.0.1

# Database
psycopg2-binary==2.9.9
sqlite3

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
scrapy==2.11.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
python-magic==0.4.27

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# AI/ML
openai==1.3.7
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.25.2
scikit-learn==1.3.2

# Text Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Utilities
click==8.1.7
tqdm==4.66.1
schedule==1.2.0
psutil==5.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# n8n Integration
n8n-python-sdk==0.1.0

# Async Support
asyncio
aiohttp==3.9.1
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

